@@ -35,119 +35,118 @@
     "schemaVersion": "3.13"
   },
   "dependencies": {
-    "@hookform/resolvers": "3.3.0",
-    "@next/env": "13.4.19",
-    "@radix-ui/react-alert-dialog": "1.0.4",
-    "@saleor/app-sdk": "0.43.0",
+    "@hookform/resolvers": "3.3.2",
+    "@next/env": "14.0.4",
+    "@radix-ui/react-alert-dialog": "1.0.5",
+    "@saleor/app-sdk": "0.47.2",
     "@saleor/macaw-ui": "0.8.0-pre.123",
-    "@sentry/nextjs": "7.77.0",
-    "@t3-oss/env-nextjs": "0.6.1",
-    "@tanstack/react-query": "4.33.0",
-    "@tanstack/react-query-devtools": "4.33.0",
-    "@trpc/client": "10.37.1",
-    "@trpc/next": "10.37.1",
-    "@trpc/react-query": "10.37.1",
-    "@trpc/server": "10.37.1",
+    "@sentry/nextjs": "7.86.0",
+    "@t3-oss/env-nextjs": "0.7.1",
+    "@tanstack/react-query": "4",
+    "@tanstack/react-query-devtools": "4.36.1",
+    "@trpc/client": "10.44.1",
+    "@trpc/next": "10.44.1",
+    "@trpc/server": "10.44.1",
     "@urql/exchange-auth": "1.0.0",
-    "@vanilla-extract/css": "1.13.0",
-    "@vanilla-extract/next-plugin": "2.3.0",
-    "@vanilla-extract/recipes": "0.5.0",
-    "@vitejs/plugin-react": "4.0.4",
+    "@vanilla-extract/css": "1.14.0",
+    "@vanilla-extract/next-plugin": "2.3.2",
+    "@vanilla-extract/recipes": "0.5.1",
+    "@vitejs/plugin-react": "4.2.1",
     "ajv": "8.12.0",
     "ajv-formats": "2.1.1",
     "bluebird": "3.7.2",
     "classnames": "2.3.2",
     "eslint-plugin-node": "11.1.0",
-    "graphql": "16.8.0",
+    "graphql": "16.8.1",
     "graphql-tag": "2.12.6",
-    "jose": "4.14.4",
-    "jsdom": "22.1.0",
+    "jose": "5.1.3",
+    "jsdom": "23.0.1",
     "lodash-es": "4.17.21",
-    "modern-errors": "6.0.0",
-    "modern-errors-http": "4.0.0",
-    "modern-errors-serialize": "5.0.0",
-    "next": "13.4.19",
+    "modern-errors": "7.0.0",
+    "modern-errors-http": "5.0.0",
+    "modern-errors-serialize": "6.0.0",
+    "next": "14.0.4",
     "omit-deep-lodash": "1.1.7",
-    "pino": "8.15.0",
-    "pino-pretty": "10.2.0",
+    "pino": "8.16.2",
+    "pino-pretty": "10.2.3",
     "react": "18.2.0",
     "react-dom": "18.2.0",
-    "react-hook-form": "7.45.4",
+    "react-hook-form": "7.49.1",
     "semver": "7.5.4",
-    "stripe": "12.16.0",
-    "stripe-event-types": "2.4.0",
-    "tsx": "3.12.7",
+    "stripe": "14.8.0",
+    "stripe-event-types": "3.1.0",
+    "tsx": "4.6.2",
     "url-join": "5.0.0",
     "urql": "3.0.4",
-    "uuidv7": "0.6.2",
-    "vite": "4.4.9",
-    "vitest": "0.34.2",
-    "webpack": "5.88.2",
-    "yup": "1.2.0",
-    "zod": "3.22.2",
-    "zustand": "4.4.1"
+    "uuidv7": "0.6.3",
+    "webpack": "5.89.0",
+    "yup": "1.3.2",
+    "zod": "3.22.4",
+    "zustand": "4.4.7"
   },
   "devDependencies": {
-    "@changesets/cli": "2.26.2",
+    "@changesets/cli": "2.27.1",
     "@graphql-codegen/add": "5.0.0",
     "@graphql-codegen/cli": "5.0.0",
     "@graphql-codegen/introspection": "4.0.0",
     "@graphql-codegen/typed-document-node": "5.0.1",
     "@graphql-codegen/typescript": "4.0.1",
     "@graphql-codegen/typescript-operations": "4.0.1",
-    "@graphql-codegen/typescript-urql": "3.7.3",
-    "@graphql-codegen/urql-introspection": "2.2.1",
+    "@graphql-codegen/typescript-urql": "4.0.0",
+    "@graphql-codegen/urql-introspection": "3.0.0",
     "@graphql-typed-document-node/core": "3.2.0",
-    "@next/env": "13.4.12",
     "@pollyjs/adapter-fetch": "6.0.6",
     "@pollyjs/adapter-node-http": "6.0.6",
     "@pollyjs/core": "6.0.6",
     "@pollyjs/persister-fs": "6.0.6",
     "@saleor/eslint-plugin-saleor-app": "0.1.2",
     "@saleor/json-schema-compiler": "0.1.2",
-    "@testing-library/jest-dom": "6.1.1",
-    "@testing-library/react": "14.0.0",
+    "@testing-library/jest-dom": "6.1.5",
+    "@testing-library/react": "14.1.2",
     "@testing-library/react-hooks": "8.0.1",
-    "@types/bluebird": "3.5.38",
-    "@types/lodash-es": "4.17.8",
-    "@types/node": "20.5.4",
-    "@types/omit-deep-lodash": "1.1.1",
-    "@types/react": "18.2.21",
-    "@types/react-dom": "18.2.7",
-    "@types/semver": "7.5.0",
-    "@types/setup-polly-jest": "0.5.2",
-    "@typescript-eslint/eslint-plugin": "6.4.1",
-    "@typescript-eslint/parser": "6.4.1",
-    "@vanilla-extract/vite-plugin": "3.9.0",
+    "@types/bluebird": "3.5.42",
+    "@types/lodash-es": "4.17.12",
+    "@types/node": "20.10.4",
+    "@types/omit-deep-lodash": "1.1.3",
+    "@types/react": "18.2.45",
+    "@types/react-dom": "18.2.17",
+    "@types/semver": "7.5.6",
+    "@types/setup-polly-jest": "0.5.5",
+    "@typescript-eslint/eslint-plugin": "6.14.0",
+    "@typescript-eslint/parser": "6.14.0",
+    "@vanilla-extract/vite-plugin": "3.9.3",
     "@vitest/coverage-v8": "0.34.2",
-    "dependency-cruiser": "13.1.4",
-    "eslint": "8.47.0",
-    "eslint-config-next": "13.4.19",
-    "eslint-config-prettier": "9.0.0",
-    "eslint-plugin-import": "2.28.1",
+    "vite": "5.0.8",
+    "dependency-cruiser": "15.5.0",
+    "eslint": "8.55.0",
+    "eslint-config-next": "14.0.4",
+    "eslint-config-prettier": "9.1.0",
+    "eslint-plugin-import": "2.29.0",
     "eslint-plugin-require-form-method": "1.0.2",
-    "eslint-plugin-testing-library": "6.0.0",
-    "eslint-plugin-vitest": "0.2.8",
+    "eslint-plugin-testing-library": "6.2.0",
+    "eslint-plugin-vitest": "0.3.16",
     "husky": "8.0.3",
-    "json-schema-to-typescript": "13.0.2",
-    "lint-staged": "14.0.1",
-    "next-test-api-route-handler": "3.1.8",
-    "prettier": "3.0.2",
+    "json-schema-to-typescript": "13.1.1",
+    "lint-staged": "15.2.0",
+    "next-test-api-route-handler": "3.1.10",
+    "prettier": "3.1.1",
     "setup-polly-jest": "0.11.0",
-    "ts-node": "10.9.1",
-    "typescript": "5.1.6",
-    "vite-tsconfig-paths": "4.2.0"
+    "ts-node": "10.9.2",
+    "typescript": "5.3.3",
+    "vite-tsconfig-paths": "4.2.2",
+    "vitest": "0.34.2"
   },
   "engines": {
     "npm": ">=8.0.0 <10.0.0",
-    "node": ">=18.0.0 <19.0.0",
-    "pnpm": ">=8.0.0 <9.0.0"
+    "node": "^20.0.0 || ^22.0.0",
+    "pnpm": "~8.12.0"
   },
-  "packageManager": "pnpm@8.6.10",
+  "packageManager": "pnpm@8.12.0",
   "pnpm": {
     "overrides": {
       "@urql/exchange-auth>@urql/core": "3.2.2",
-      "@typescript-eslint/parser": "$@typescript-eslint/parser"
+      "@typescript-eslint/parser": "$@typescript-eslint/parser",
+      "vite": "$vite"
     }
   }
 }
