@@ -1,5 +1,5 @@
 /** @type {import('dependency-cruiser').IConfiguration} */
-module.exports = {
+export default {
   forbidden: [
     /* rules from the 'recommended' preset: */
     {
@@ -96,7 +96,9 @@ module.exports = {
         "This module depends on a module that cannot be found ('resolved to disk'). If it's an npm " +
         "module: add it to your package.json. In all other cases you likely already know what to do.",
       severity: "error",
-      from: {},
+      from: {
+        pathNot: "generated/graphql.ts",
+      },
       to: {
         pathNot: "@saleor/app-sdk/types",
         couldNotResolve: true,
